eliminate possibly-ambiguous override of SWAP
git-svn-id: svn://tug.org/texlive/trunk/Build/source@13650 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/web2c/xetexdir/XeTeXFontMgr.cpp
@@ -584,7 +584,7 @@ XeTeXFontMgr::getOpSizeRecAndStyleFlags(Font* theFont)
 
 		const POSTTable* postTable = (const POSTTable*)getFontTablePtr(font, LE_POST_TABLE_TAG);
 		if (postTable != NULL) {
-			theFont->slant = (int)(1000 * (tan(Fix2X(-SWAP(postTable->italicAngle)) * M_PI / 180.0)));
+			theFont->slant = (int)(1000 * (tan(Fix2X(-SWAP(UInt32(postTable->italicAngle))) * M_PI / 180.0)));
 		}
 		deleteFont(font);
 	}
